refactor: some fixes
