+++ .gitignore
@@ -4,3 +4,4 @@ documentation
 build
 ./documentation
 /documentation/
+logs
